Rename templates dir to template.
