+++ pkg/template/sample/report/BUILD
@@ -12,14 +12,14 @@ go_proto_compile(
         "mixer/v1/config/descriptor/value_type.proto": "istio.io/api/mixer/v1/config/descriptor",
         "google/protobuf/descriptor.proto": "github.com/golang/protobuf/protoc-gen-go/descriptor",
         "google/protobuf/struct.proto": "github.com/golang/protobuf/ptypes/struct",
-        "pkg/templates/mixer/TemplateExtensions.proto": "istio.io/mixer/pkg/templates/mixer",
+        "pkg/template/mixer/TemplateExtensions.proto": "istio.io/mixer/pkg/template/mixer",
     },
     imports = [
         "external/com_github_google_protobuf/src",
         "external/com_github_istio_api",
     ],
     inputs = [
-        "//pkg/templates/mixer:protos",
+        "//pkg/template/mixer:protos",
         "@com_github_google_protobuf//:well_known_protos",
         "@com_github_istio_api//:mixer/v1/config/descriptor_protos",
     ],
@@ -43,7 +43,7 @@ go_library(
     ],
     deps = [
         "//pkg/adapter/config:go_default_library",
-        "//pkg/templates/mixer:go_default_library",
+        "//pkg/template/mixer:go_default_library",
         "@com_github_golang_protobuf//proto:go_default_library",
         "@com_github_golang_protobuf//protoc-gen-go/descriptor:go_default_library",
         "@com_github_istio_api//:mixer/v1/config/descriptor",
