+++ pkg/template/sample/report/ReportTesterTemplate.gen.proto
@@ -3,12 +3,17 @@ syntax = "proto3";
 package istio.mixer.adapter.sample.report;
 
 import "mixer/v1/config/descriptor/value_type.proto";
-import "mixer/pkg/templates/mixer/TemplateExtensions.proto";
+import "pkg/template/mixer/TemplateExtensions.proto";
 
 option (istio.mixer.v1.config.template.template_variety) = TEMPLATE_VARIETY_REPORT;
 option (istio.mixer.v1.config.template.template_name) = "Sample";
 
-message Template {
+message Type {
     istio.mixer.v1.config.descriptor.ValueType value = 2;
     map<string, istio.mixer.v1.config.descriptor.ValueType> dimensions = 3;
+}
+
+message ConstructorParam {
+    string value = 2;
+    map<string, string> dimensions = 3;
 }
