+++ pkg/template/sample/report/ReportTesterTemplate.proto
@@ -3,17 +3,12 @@ syntax = "proto3";
 package istio.mixer.adapter.sample.report;
 
 import "mixer/v1/config/descriptor/value_type.proto";
-import "pkg/templates/mixer/TemplateExtensions.proto";
+import "mixer/pkg/template/mixer/TemplateExtensions.proto";
 
 option (istio.mixer.v1.config.template.template_variety) = TEMPLATE_VARIETY_REPORT;
 option (istio.mixer.v1.config.template.template_name) = "Sample";
 
-message Type {
+message Template {
     istio.mixer.v1.config.descriptor.ValueType value = 2;
     map<string, istio.mixer.v1.config.descriptor.ValueType> dimensions = 3;
-}
-
-message ConstructorParam {
-    string value = 2;
-    map<string, string> dimensions = 3;
 }
